+++ sqlobject/col.py
@@ -1505,7 +1505,11 @@ class TimeValidator(DateTimeValidator):
                 raise validators.Invalid(
                     "the value for the TimeCol '%s' must has days=0, "
                     "it has days=%d" % (self.name, value.days), value, state)
-            return datetime.time(*time.gmtime(value.seconds)[3:6])
+            print("[DEBUG1]:", value.microseconds)
+            return datetime.time(
+                *time.gmtime(value.seconds)[3:6],
+                microsecond=value.microseconds
+            )
         value = super(TimeValidator, self).to_python(value, state)
         if isinstance(value, datetime.datetime):
             value = value.time()