+++ sqlobject/mysql/mysqlconnection.py
@@ -1,5 +1,3 @@
-import os
-
 from sqlobject import col, dberrors
 from sqlobject.compat import PY2
 from sqlobject.dbconnection import DBAPI
@@ -482,9 +480,6 @@ class MySQLConnection(DBAPI):
     def can_use_microseconds(self):
         if self._can_use_microseconds is not None:
             return self._can_use_microseconds
-        if os.environ.get('APPVEYOR') or os.environ.get('TRAVIS'):
-            self._can_use_microseconds = False
-            return False
         server_version = self.server_version()
         if server_version is None:
             return None