+++ sqlobject/tests/dbtest.py
@@ -120,11 +120,7 @@ except Exception as e:
     if 'sphinx' not in sys.modules:
         print("Could not open database: %s" % e, file=sys.stderr)
 else:
-    if (connection.dbName == 'firebird') \
-        or (
-            (connection.dbName == 'mysql')
-            and ((os.environ.get('APPVEYOR')) or (os.environ.get('TRAVIS')))
-    ):
+    if (connection.dbName == 'firebird'):
         use_microseconds(False)